Petite maj pour afficher des truc, faudrait revoir lm'architecture du logiciel je pense...

+++ taupe_models/src/Point.java
@@ -1,25 +1,15 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-
-
 /**
  *
  * @author flo
  */
 public class Point {
-    int x;
-    int y;
-    int z;
+    float x;
+    float y;
+    float z;
     
-    public Point(int x, int y, int z){
+    public Point(float x, float y, float z){
         this.x = x;
         this.y = y;
         this.z = z;
-    }
-    
-    
+    }    
 }
